+++ NEWS.md
@@ -1,8 +1,11 @@
 # bvarnet 1.0.1.9000
+
 * Minor bug and documentation fixes.
 * Introduce toolchain free installation via `bvarnet_setup_models()`.
+* Change default CI to 95% and introduce argument to vary CI width.
 
 # bvarnet 1.0.1
+
 * Add installation safeguard on CRAN.
 
 # bvarnet 1.0.0